+++ lib/src/oauth_web_screen.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:oauth2/oauth2.dart';
+import 'package:http/http.dart' as http;
 import 'package:oauth_webauth/oauth_webauth.dart';
 
 class OAuthWebScreen extends StatelessWidget {
@@ -14,6 +15,9 @@ class OAuthWebScreen extends StatelessWidget {
     final String? baseUrl,
     required String clientId,
     String? clientSecret,
+    String? delimiter,
+    bool? basicAuth,
+    http.Client? httpClient,
     List<String>? scopes,
     String? loginHint,
     List<String>? promptValues,
@@ -48,6 +52,9 @@ class OAuthWebScreen extends StatelessWidget {
           baseUrl: baseUrl,
           clientId: clientId,
           clientSecret: clientSecret,
+          delimiter: delimiter,
+          basicAuth: basicAuth,
+          httpClient: httpClient,
           scopes: scopes,
           loginHint: loginHint,
           promptValues: promptValues,
@@ -68,6 +75,9 @@ class OAuthWebScreen extends StatelessWidget {
                   baseUrl: baseUrl,
                   clientId: clientId,
                   clientSecret: clientSecret,
+                  delimiter: delimiter,
+                  basicAuth: basicAuth,
+                  httpClient: httpClient,
                   scopes: scopes,
                   loginHint: loginHint,
                   promptValues: promptValues,
@@ -88,14 +98,74 @@ class OAuthWebScreen extends StatelessWidget {
                 )));
   }
 
+  /// A URL provided by the authorization server that serves as the base for the
+  /// URL that the resource owner will be redirected to to authorize this
+  /// client.
+  ///
+  /// This will usually be listed in the authorization server's OAuth2 API
+  /// documentation.
   final String authorizationEndpointUrl;
+
+  /// A URL provided by the authorization server that this library uses to
+  /// obtain long-lasting credentials.
+  ///
+  /// This will usually be listed in the authorization server's OAuth2 API
+  /// documentation.
   final String tokenEndpointUrl;
+
+  /// The redirectUrl to which the authorization server will redirect to when
+  /// authorization flow completes.
   final String redirectUrl;
+
+  /// It is another redirectUrl to be used by the authorization server when the
+  /// user decides to go back to application instead of completing
+  /// authorization flow.
   final String? baseUrl;
+
+  /// The client identifier for this client.
+  ///
+  /// The authorization server will issue each client a separate client
+  /// identifier and secret, which allows the server to tell which client is
+  /// accessing it. Some servers may also have an anonymous identifier/secret
+  /// pair that any client may use.
+  ///
+  /// This is usually global to the program using this library.
   final String clientId;
+
+  /// The client secret for this client.
+  ///
+  /// The authorization server will issue each client a separate client
+  /// identifier and secret, which allows the server to tell which client is
+  /// accessing it. Some servers may also have an anonymous identifier/secret
+  /// pair that any client may use.
+  ///
+  /// This is usually global to the program using this library.
+  ///
+  /// Note that clients whose source code or binary executable is readily
+  /// available may not be able to make sure the client secret is kept a secret.
+  /// This is fine; OAuth2 servers generally won't rely on knowing with
+  /// certainty that a client is who it claims to be.
   final String? clientSecret;
+
+  /// A [String] used to separate scopes; defaults to `" "`.
+  final String? delimiter;
+
+  /// Whether to use HTTP Basic authentication for authorizing the client.
+  final bool? basicAuth;
+
+  /// The HTTP client used to make HTTP requests.
+  final http.Client? httpClient;
+
+  /// The scopes that the client is requesting access to.
   final List<String>? scopes;
+
+  /// /// Hint to the Authorization Server about the login identifier the End-User might use to log in.
   final String? loginHint;
+
+  /// List of ASCII string values that specifies whether the Authorization Server prompts the End-User for re-authentication and consent.
+  /// e.g: promptValues: ['login'].
+  /// In this case it prompts the user login even if they have already signed in.
+  /// The list of supported values depends on the identity provider.
   final List<String>? promptValues;
 
   /// This function will be called when user successfully authenticates.
@@ -147,6 +217,9 @@ class OAuthWebScreen extends StatelessWidget {
     this.baseUrl,
     required this.clientId,
     this.clientSecret,
+    this.delimiter,
+    this.basicAuth,
+    this.httpClient,
     this.scopes,
     this.loginHint,
     this.promptValues,
@@ -185,6 +258,9 @@ class OAuthWebScreen extends StatelessWidget {
                 tokenEndpointUrl: tokenEndpointUrl,
                 clientId: clientId,
                 clientSecret: clientSecret,
+                delimiter: delimiter,
+                basicAuth: basicAuth,
+                httpClient: httpClient,
                 redirectUrl: redirectUrl,
                 baseUrl: baseUrl,
                 scopes: scopes,